Remove unnecessary files and rename
