+++ solutions/standing-ovation/go/approach3.go
