--- solutions/standing-ovation/java/A-small-practice.in
@@ -1,101 +1,0 @@
-100
-6 4069616
-1 01
-6 0000001
-4 11111
-6 5050001
-6 1111111
-1 88
-1 09
-6 6000861
-6 0030702
-6 8535649
-6 2851658
-6 0000001
-2 004
-1 41
-6 5641739
-6 1011021
-6 0008001
-6 0000001
-6 4060001
-6 0000002
-6 0896543
-6 8835525
-6 5000072
-5 800001
-0 1
-6 3010001
-1 11
-0 3
-6 0000001
-6 2102121
-6 0080104
-6 1111002
-5 002509
-2 201
-6 8000001
-3 1021
-3 1122
-1 06
-4 00009
-4 20122
-3 8241
-6 6000081
-6 2050001
-4 00801
-6 2222222
-6 2110001
-6 8189804
-6 0000001
-6 0780301
-0 1
-1 94
-6 0000001
-6 0800003
-5 507001
-1 62
-2 001
-6 9999999
-6 0656715
-4 00021
-6 6397845
-2 751
-5 005005
-5 350901
-4 02021
-6 2111001
-6 0471798
-4 80001
-3 9504
-0 1
-3 0011
-6 3800001
-6 0000001
-5 120701
-6 0122111
-6 4089766
-1 01
-6 0840001
-6 3004001
-5 846687
-0 1
-5 290021
-5 110011
-6 0800001
-0 1
-6 0000001
-1 97
-0 1
-5 000201
-6 6000001
-6 6263018
-2 301
-6 6801701
-6 2101201
-1 01
-6 2212121
-6 2222112
-6 0000001
-3 0071
-6 0000001
